+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# Changelog
+
+## February 9 2017
+
+ * Add "format" attribute to `wpDate` in `wpPosts`
+ * Change "format_out" to just "format" in `wpCustomDate`
+
+## Old changelog
+
 * 0.1.1.2 - Add missing Misc module for test suite, expose modules for test suite.
 * 0.1.1.1 - Update for GHC 7.10.
 * 0.1.1.0 - Initial release.